requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
jinja2
qdrant-client
sentence-transformers
httpx
python-dotenv
huggingface_hub